js/xpconnect/tests/mochitest/test_bug809674.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=809674
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 809674</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=809674">Mozilla Bug 809674</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 809674 **/

var mse = document.createEvent('TimeEvents');
mse.initTimeEvent('', {}, '');
try {
  mse.view.onunload;
  ok(false, "Should have thrown");
} catch(e) {
  ok(!!/implicit_jscontext/.exec(e), "Should throw a helpful error");
}

</script>
</pre>
</body>
</html>
